src/bk-login/requirements.txt
--index-url https://mirrors.tencent.com/pypi/simple

annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "3.12"
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "3.12"
bk-crypto-python-sdk==2.0.0 ; python_version >= "3.11" and python_version < "3.12"
bk-notice-sdk==1.3.2 ; python_version >= "3.11" and python_version < "3.12"
bkapi-client-core==1.2.0 ; python_version >= "3.11" and python_version < "3.12"
blue-krill==2.0.7 ; python_version >= "3.11" and python_version < "3.12"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "3.12"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "3.12"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.12" and platform_system == "Windows"
concurrent-log-handler==0.9.25 ; python_version >= "3.11" and python_version < "3.12"
cryptography==44.0.0 ; python_version >= "3.11" and python_version < "3.12"
curlify==2.2.1 ; python_version >= "3.11" and python_version < "3.12"
dacite==1.8.1 ; python_version >= "3.11" and python_version < "3.12"
deprecated==1.2.15 ; python_version >= "3.11" and python_version < "3.12"
django-cors-headers==4.6.0 ; python_version >= "3.11" and python_version < "3.12"
django-environ==0.11.2 ; python_version >= "3.11" and python_version < "3.12"
django-prometheus==2.3.1 ; python_version >= "3.11" and python_version < "3.12"
django==4.2.16 ; python_version >= "3.11" and python_version < "3.12"
gevent==24.11.1 ; python_version >= "3.11" and python_version < "3.12"
googleapis-common-protos==1.66.0 ; python_version >= "3.11" and python_version < "3.12"
greenlet==3.1.1 ; python_version >= "3.11" and python_version < "3.12" and platform_python_implementation == "CPython"
grpcio==1.68.1 ; python_version >= "3.11" and python_version < "3.12"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "3.12"
idna==3.10 ; python_version >= "3.11" and python_version < "3.12"
importlib-metadata==7.1.0 ; python_version >= "3.11" and python_version < "3.12"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-api==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-exporter-otlp-proto-common==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-exporter-otlp-proto-grpc==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-exporter-otlp-proto-http==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-exporter-otlp==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-celery==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-dbapi==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-django==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-logging==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-redis==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-requests==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation-wsgi==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-instrumentation==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-proto==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-sdk==1.25.0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-semantic-conventions==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
opentelemetry-util-http==0.46b0 ; python_version >= "3.11" and python_version < "3.12"
packaging==24.2 ; python_version >= "3.11" and python_version < "3.12"
portalocker==3.0.0 ; python_version >= "3.11" and python_version < "3.12"
prometheus-client==0.21.1 ; python_version >= "3.11" and python_version < "3.12"
protobuf==4.25.5 ; python_version >= "3.11" and python_version < "3.12"
pycparser==2.22 ; python_version >= "3.11" and python_version < "3.12"
pycryptodomex==3.21.0 ; python_version >= "3.11" and python_version < "3.12"
pydantic-core==2.16.3 ; python_version >= "3.11" and python_version < "3.12"
pydantic==2.6.4 ; python_version >= "3.11" and python_version < "3.12"
pyjwt==2.10.1 ; python_version >= "3.11" and python_version < "3.12"
pyjwt[cryptography]==2.10.1 ; python_version >= "3.11" and python_version < "3.12"
pymysql==1.1.1 ; python_version >= "3.11" and python_version < "3.12"
python-editor==1.0.4 ; python_version >= "3.11" and python_version < "3.12"
python-json-logger==2.0.7 ; python_version >= "3.11" and python_version < "3.12"
pywin32==308 ; python_version >= "3.11" and python_version < "3.12" and platform_system == "Windows"
requests==2.32.3 ; python_version >= "3.11" and python_version < "3.12"
sentry-sdk==1.45.0 ; python_version >= "3.11" and python_version < "3.12"
setuptools==75.6.0 ; python_version >= "3.11" and python_version < "3.12"
six==1.17.0 ; python_version >= "3.11" and python_version < "3.12"
sqlparse==0.5.2 ; python_version >= "3.11" and python_version < "3.12"
toml==0.10.2 ; python_version >= "3.11" and python_version < "3.12"
tongsuopy-crayon==1.0.2b6 ; python_version >= "3.11" and python_version < "3.12"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "3.12"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "3.12" and sys_platform == "win32"
urllib3==1.26.20 ; python_version >= "3.11" and python_version < "3.12"
watchdog==6.0.0 ; python_version >= "3.11" and python_version < "3.12"
werkzeug==3.0.6 ; python_version >= "3.11" and python_version < "3.12"
whitenoise==6.8.2 ; python_version >= "3.11" and python_version < "3.12"
wrapt==1.17.0 ; python_version >= "3.11" and python_version < "3.12"
zipp==3.21.0 ; python_version >= "3.11" and python_version < "3.12"
zope-event==5.0 ; python_version >= "3.11" and python_version < "3.12"
zope-interface==7.2 ; python_version >= "3.11" and python_version < "3.12"
